requirements.txt
pygithub
langchain
langchain-community
openai
tiktoken
pinecone-client
langchain-pinecone
sentence-transformers
streamlit